requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Flask-Cors==4.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
PyJWT==2.8.0
python-dotenv==1.0.0
cohere==4.27
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==0.8.11
numpy==1.25.2
Pillow==10.0.0
requests==2.31.0
blinker==1.6.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
alembic==1.12.0
email-validator==2.0.0.post2
gunicorn==21.2.0